requirements.txt
aiohttp==3.10.5 
pydantic==2.9.0 
pycryptodome==3.20.0 